move dao-related stuff to dao.ts from cosmos.ts
